requirements.txt
python-dotenv==1.0.1
requests==2.32.3
pymongo==4.8.0
urllib3==2.2.2
